chore(refactoring): move plugin-init functions into a common util module
