+++ src/plugin_runtime/mod.rs
@@ -4,5 +4,18 @@ pub mod graph;
 pub mod kernel;
 pub mod resolver;
 pub mod starter;
+pub mod util;
 
 pub use self::kernel::{Error, Kernel};
+
+use crate::plugin_support::{Plugin, PluginStep};
+
+pub type PluginId = usize;
+
+pub type Injection = (Plugin, InjectionTarget);
+
+#[derive(Copy, Clone, Debug)]
+pub enum InjectionTarget {
+    BeforeStep(PluginStep),
+    AfterStep(PluginStep),
+}
